requirements.txt
dash == 2.18.2
dash-bootstrap-components == 1.7.1
numpy == 2.3.3
plotly-express == 0.4.1
pandas == 2.2.3